+++ docs/docs/auto-docs/shared-components/Recurrence/CustomRecurrenceModal/interfaces/InterfaceCustomRecurrenceModalProps.md
@@ -4,7 +4,7 @@
 
 # Interface: InterfaceCustomRecurrenceModalProps
 
-Defined in: [src/shared-components/Recurrence/CustomRecurrenceModal.tsx:28](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/Recurrence/CustomRecurrenceModal.tsx#L28)
+Defined in: [src/shared-components/Recurrence/CustomRecurrenceModal.tsx:27](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/Recurrence/CustomRecurrenceModal.tsx#L27)
 
 Props interface for the CustomRecurrenceModal component
 
@@ -14,7 +14,7 @@ Props interface for the CustomRecurrenceModal component
 
 > **customRecurrenceModalIsOpen**: `boolean`
 
-Defined in: [src/shared-components/Recurrence/CustomRecurrenceModal.tsx:40](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/Recurrence/CustomRecurrenceModal.tsx#L40)
+Defined in: [src/shared-components/Recurrence/CustomRecurrenceModal.tsx:39](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/Recurrence/CustomRecurrenceModal.tsx#L39)
 
 Whether the custom recurrence modal is open
 
@@ -24,7 +24,7 @@ Whether the custom recurrence modal is open
 
 > **endDate**: `Date`
 
-Defined in: [src/shared-components/Recurrence/CustomRecurrenceModal.tsx:36](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/Recurrence/CustomRecurrenceModal.tsx#L36)
+Defined in: [src/shared-components/Recurrence/CustomRecurrenceModal.tsx:35](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/Recurrence/CustomRecurrenceModal.tsx#L35)
 
 Event end date
 
@@ -34,7 +34,7 @@ Event end date
 
 > **hideCustomRecurrenceModal**: () => `void`
 
-Defined in: [src/shared-components/Recurrence/CustomRecurrenceModal.tsx:42](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/Recurrence/CustomRecurrenceModal.tsx#L42)
+Defined in: [src/shared-components/Recurrence/CustomRecurrenceModal.tsx:41](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/Recurrence/CustomRecurrenceModal.tsx#L41)
 
 Function to hide the custom recurrence modal
 
@@ -48,7 +48,7 @@ Function to hide the custom recurrence modal
 
 > **recurrenceRuleState**: [`InterfaceRecurrenceRule`](../../../../utils/recurrenceUtils/recurrenceTypes/interfaces/InterfaceRecurrenceRule.md)
 
-Defined in: [src/shared-components/Recurrence/CustomRecurrenceModal.tsx:30](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/Recurrence/CustomRecurrenceModal.tsx#L30)
+Defined in: [src/shared-components/Recurrence/CustomRecurrenceModal.tsx:29](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/Recurrence/CustomRecurrenceModal.tsx#L29)
 
 Current recurrence rule state
 
@@ -58,7 +58,7 @@ Current recurrence rule state
 
 > **setCustomRecurrenceModalIsOpen**: (`state`) => `void`
 
-Defined in: [src/shared-components/Recurrence/CustomRecurrenceModal.tsx:44](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/Recurrence/CustomRecurrenceModal.tsx#L44)
+Defined in: [src/shared-components/Recurrence/CustomRecurrenceModal.tsx:43](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/Recurrence/CustomRecurrenceModal.tsx#L43)
 
 Function to set custom recurrence modal open state
 
@@ -78,7 +78,7 @@ Function to set custom recurrence modal open state
 
 > **setEndDate**: (`state`) => `void`
 
-Defined in: [src/shared-components/Recurrence/CustomRecurrenceModal.tsx:38](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/Recurrence/CustomRecurrenceModal.tsx#L38)
+Defined in: [src/shared-components/Recurrence/CustomRecurrenceModal.tsx:37](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/Recurrence/CustomRecurrenceModal.tsx#L37)
 
 Function to set event end date
 
@@ -98,7 +98,7 @@ Function to set event end date
 
 > **setRecurrenceRuleState**: (`state`) => `void`
 
-Defined in: [src/shared-components/Recurrence/CustomRecurrenceModal.tsx:32](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/Recurrence/CustomRecurrenceModal.tsx#L32)
+Defined in: [src/shared-components/Recurrence/CustomRecurrenceModal.tsx:31](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/Recurrence/CustomRecurrenceModal.tsx#L31)
 
 Function to update recurrence rule state
 
@@ -118,7 +118,7 @@ Function to update recurrence rule state
 
 > **startDate**: `Date`
 
-Defined in: [src/shared-components/Recurrence/CustomRecurrenceModal.tsx:50](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/Recurrence/CustomRecurrenceModal.tsx#L50)
+Defined in: [src/shared-components/Recurrence/CustomRecurrenceModal.tsx:49](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/Recurrence/CustomRecurrenceModal.tsx#L49)
 
 Event start date
 
@@ -128,7 +128,7 @@ Event start date
 
 > **t**: (`key`) => `string`
 
-Defined in: [src/shared-components/Recurrence/CustomRecurrenceModal.tsx:48](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/Recurrence/CustomRecurrenceModal.tsx#L48)
+Defined in: [src/shared-components/Recurrence/CustomRecurrenceModal.tsx:47](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/Recurrence/CustomRecurrenceModal.tsx#L47)
 
 Translation function
