+++ docs/docs/auto-docs/shared-components/Recurrence/RecurrenceEndOptionsSection/variables/RecurrenceEndOptionsSection.md
@@ -1,0 +1,11 @@
+[Admin Docs](/)
+
+***
+
+# Variable: RecurrenceEndOptionsSection
+
+> `const` **RecurrenceEndOptionsSection**: `React.FC`\<`InterfaceRecurrenceEndOptionsSectionProps`\>
+
+Defined in: [src/shared-components/Recurrence/RecurrenceEndOptionsSection.tsx:35](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/Recurrence/RecurrenceEndOptionsSection.tsx#L35)
+
+Recurrence end options section (never, on date, after count)
