+++ docs/docs/auto-docs/shared-components/Recurrence/RecurrenceFrequencySection/variables/RecurrenceFrequencySection.md
@@ -1,0 +1,11 @@
+[Admin Docs](/)
+
+***
+
+# Variable: RecurrenceFrequencySection
+
+> `const` **RecurrenceFrequencySection**: `React.FC`\<`InterfaceRecurrenceFrequencySectionProps`\>
+
+Defined in: [src/shared-components/Recurrence/RecurrenceFrequencySection.tsx:17](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/Recurrence/RecurrenceFrequencySection.tsx#L17)
+
+Frequency and interval selection section
