+++ docs/docs/auto-docs/shared-components/Recurrence/RecurrenceMonthlySection/variables/RecurrenceMonthlySection.md
@@ -1,0 +1,11 @@
+[Admin Docs](/)
+
+***
+
+# Variable: RecurrenceMonthlySection
+
+> `const` **RecurrenceMonthlySection**: `React.FC`\<`InterfaceRecurrenceMonthlySectionProps`\>
+
+Defined in: [src/shared-components/Recurrence/RecurrenceMonthlySection.tsx:19](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/Recurrence/RecurrenceMonthlySection.tsx#L19)
+
+Monthly recurrence options section
