+++ docs/docs/auto-docs/shared-components/Recurrence/RecurrenceWeeklySection/variables/RecurrenceWeeklySection.md
@@ -1,0 +1,11 @@
+[Admin Docs](/)
+
+***
+
+# Variable: RecurrenceWeeklySection
+
+> `const` **RecurrenceWeeklySection**: `React.FC`\<`InterfaceRecurrenceWeeklySectionProps`\>
+
+Defined in: [src/shared-components/Recurrence/RecurrenceWeeklySection.tsx:24](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/Recurrence/RecurrenceWeeklySection.tsx#L24)
+
+Weekly recurrence day selection section
