+++ docs/docs/auto-docs/shared-components/Recurrence/RecurrenceYearlySection/variables/RecurrenceYearlySection.md
@@ -1,0 +1,11 @@
+[Admin Docs](/)
+
+***
+
+# Variable: RecurrenceYearlySection
+
+> `const` **RecurrenceYearlySection**: `React.FC`\<`InterfaceRecurrenceYearlySectionProps`\>
+
+Defined in: [src/shared-components/Recurrence/RecurrenceYearlySection.tsx:13](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/shared-components/Recurrence/RecurrenceYearlySection.tsx#L13)
+
+Yearly recurrence options section
