+++ docs/docs/auto-docs/utils/recurrenceUtils/recurrenceUtilityFunctions/functions/getDayName.md
@@ -1,0 +1,25 @@
+[Admin Docs](/)
+
+***
+
+# Function: getDayName()
+
+> **getDayName**(`dayIndex`): `string`
+
+Defined in: [src/utils/recurrenceUtils/recurrenceUtilityFunctions.ts:345](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/utils/recurrenceUtils/recurrenceUtilityFunctions.ts#L345)
+
+Gets the full day name from a day index
+
+## Parameters
+
+### dayIndex
+
+`number`
+
+The day index (0-6, where 0 is Sunday)
+
+## Returns
+
+`string`
+
+The full day name
