+++ docs/docs/auto-docs/utils/recurrenceUtils/recurrenceUtilityFunctions/functions/getOrdinalString.md
@@ -1,0 +1,25 @@
+[Admin Docs](/)
+
+***
+
+# Function: getOrdinalString()
+
+> **getOrdinalString**(`num`): `string`
+
+Defined in: [src/utils/recurrenceUtils/recurrenceUtilityFunctions.ts:335](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/utils/recurrenceUtils/recurrenceUtilityFunctions.ts#L335)
+
+Converts a number to its ordinal string representation
+
+## Parameters
+
+### num
+
+`number`
+
+The number to convert (1-5)
+
+## Returns
+
+`string`
+
+The ordinal string (e.g., "first", "second", etc.)
