+++ docs/docs/auto-docs/utils/recurrenceUtils/recurrenceUtilityFunctions/functions/getWeekOfMonth.md
@@ -1,0 +1,25 @@
+[Admin Docs](/)
+
+***
+
+# Function: getWeekOfMonth()
+
+> **getWeekOfMonth**(`date`): `number`
+
+Defined in: [src/utils/recurrenceUtils/recurrenceUtilityFunctions.ts:324](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/utils/recurrenceUtils/recurrenceUtilityFunctions.ts#L324)
+
+Calculates which week of the month a given date falls in
+
+## Parameters
+
+### date
+
+`Date`
+
+The date to calculate the week for
+
+## Returns
+
+`number`
+
+The week number (1-5) within the month
